+++ test/unit/sonarqube_code_coverage.sh
@@ -10,7 +10,6 @@ echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/archive_email.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/capture_email.py
-coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/check_nonprocess.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/connect_process.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/connect_rmq.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/convert_bytes.py
@@ -19,13 +18,13 @@ coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/get_text.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/help_message.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/load_cfg.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/main.py
-coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/parse_email.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/process_attach.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/process_file.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/process_from.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/process_message.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/process_subj.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/pub_to_rmq.py
+coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/read_email.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/run_program.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/process_debug.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/archive_email_debug.py